pyproject.toml
[tool.poetry]
name        = "spectralmc"
version     = "0.0.1"
description = "GPU-accelerated library for SDE simulation"
authors     = ["tuee22"]

packages = [
  { include = "spectralmc", from = "src" }
]

# --------------------------------------------------------------------------- #
#        P Y T O R C H   C U D A   1 2 . 8   W H E E L   S O U R C E          #
# --------------------------------------------------------------------------- #
[[tool.poetry.source]]
name     = "pytorch-cu128"
url      = "https://download.pytorch.org/whl/cu128"
priority = "explicit"   # only used when a dependency asks for it


# --------------------------------------------------------------------------- #
#                       R U N T I M E   D E P E N D E N C I E S               #
# --------------------------------------------------------------------------- #
[tool.poetry.dependencies]
python = ">=3.12,<3.13"

# core
numpy          = ">=2.0,<3.0"
pandas         = ">=2.0,<3.0"
pandas-stubs   = ">=2.2.3,<3.0"
polars         = ">=1.26,<2.0"
pandera        = { version = ">=0.23.1,<1.0.0", extras = ["mypy","dask","polars"] }
aiohttp        = ">=3.8,<4.0"
aiofiles       = ">=24.1,<25.0"
types-aiofiles = ">=24.1,<25.0"

# distributed compute
dask = { version = ">=2025.3,<2026.0", extras = ["complete"] }
ray  = ">=2.44,<3.0"

# data clients
pulsar-client = ">=3.6,<4.0"
minio         = ">=7.2,<8.0"
protobuf      = ">=6.30,<7.0"
boto3         = ">=1.34.70,<2.0.0"
boto3-stubs   = ">=1.34.70,<2.0.0"
aioboto3      = ">=13.0,<14.0"
types-aiobotocore = { version = ">=2.13,<3.0", extras = ["s3"] }

# numerics
scipy            = ">=1.13,<2.0"
scikit-learn     = ">=1.6,<2.0"

# --- PyTorch 2.7.0 compiled for CUDA 12.8 (sm_120) ------------------------- #
torch            = { version = "2.7.0+cu128", source = "pytorch-cu128" }
safetensors      = ">=0.5.3,<1.0"
stable_baselines3 = ">=2.6,<3.0"
cupy-cuda12x     = ">=13.4,<14.0"
numba            = ">=0.61,<1.0"
pystan           = ">=3.10,<4.0"
pymc             = ">=5.22,<6.0"
statsmodels      = ">=0.14.4,<1.0"
QuantLib         = ">=1.37,<2.0"

# Jupyter / viz
jupyterlab  = ">=3.0,<4.0"
plotly      = ">=6.0.1,<7.0"
matplotlib  = ">=3.10.1,<4.0"
seaborn     = ">=0.13.2,<1.0"
tensorboard = ">=2.19.0,<3.0"

# runtime typing / settings
pydantic           = { version = ">=2.9,<3.0", extras = ["mypy"] }
pydantic_settings  = ">=2.7,<3.0"


# --------------------------------------------------------------------------- #
#                       D E V   &   T E S T   D E P S                         #
# --------------------------------------------------------------------------- #
[tool.poetry.group.dev.dependencies]
pytest        = ">=8.2,<9.0"
pytest-cov    = ">=5.0,<6.0"
hypothesis    = ">=6.102,<7.0"

# static analysis / formatting
mypy   = ">=1.15,<2.0"
black  = ">=25.1,<26.0"

# --------------------------------------------------------------------------- #
#                            S C R I P T S                                    #
# --------------------------------------------------------------------------- #
pytest-asyncio = "^1.2.0"
[tool.poetry.scripts]
test-all = "spectralmc.test_runner:run_all_tests"

# --------------------------------------------------------------------------- #
#                               B U I L D                                     #
# --------------------------------------------------------------------------- #
[build-system]
requires      = ["poetry-core>=1.1.0"]
build-backend = "poetry.core.masonry.api"


# --------------------------------------------------------------------------- #
#                       T O O L   C O N F I G U R A T I O N                   #
# --------------------------------------------------------------------------- #
[tool.mypy]
mypy_path = "stubs"
python_version = "3.12"
files = ["src/spectralmc", "tests", "examples"]
exclude = ["src/spectralmc/proto/.*_pb2\\.py$"]
plugins = ["pydantic.mypy"]

# Core strictness (equivalent to --strict --disallow-any-explicit)
strict = true
disallow_any_explicit = true

# Additional rigorous checks beyond --strict
disallow_any_unimported = true       # No Any from missing import stubs
disallow_any_decorated = true        # No Any from untyped decorators
warn_unreachable = true              # Warn about unreachable code
extra_checks = true                  # Enable newer experimental checks

# Disable warn_unreachable for tests (runtime assertions valuable even if statically guaranteed)
[[tool.mypy.overrides]]
module = "tests.*"
warn_unreachable = false

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[tool.pytest.ini_options]
testpaths = "tests"
addopts   = "-ra -q -m 'not gpu'"
asyncio_mode = "auto"
markers   = [
    "gpu: needs CUDA",
    "asyncio: async test"
]
filterwarnings = [
    # Hide SWIG-generated DeprecationWarnings raised while importing QuantLib
    "ignore::DeprecationWarning:.*QuantLib.*",
    # Hide IFFT imaginary component warnings in tests (expected for untrained models)
    "ignore:IFFT imaginary component.*exceeds tolerance:RuntimeWarning",
]